+++ include/simInterfaceTools.h
@@ -2,5 +2,5 @@
 #include <fstream>
 #include <opencv2/opencv.hpp>
 
-void writeToKvecFile(cv::Point2i kvecs, double amplitude, double phase,  bool isFinal);
+void writeToKvecFile(cv::Point2d kvecs, double amplitude, double phase,  bool isFinal);
 void writeDoneKvecFile();
